feature(Modal): added service to handle modal views
